+++ Backend/internal/schemas/invetory_item_schema.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from datetime import datetime
 from pydantic import BaseModel
 
@@ -7,7 +8,7 @@ class InventoryItemBase(BaseModel):
     section_id: int | None = None
     description: str | None = None
     has_critical_stock: bool = False
-    critical_stock_quantity: int | None = None
+    critical_stock_quantity: Decimal | None = None
     comments: str | None = None
     is_deleted: bool = False
 
@@ -17,7 +18,7 @@ class InventoryItemCreate(BaseModel):
     section_id: int | None = None
     description: str | None = None
     has_critical_stock: bool = False
-    critical_stock_quantity: int | None = None
+    critical_stock_quantity: Decimal | None = None
     comments: str | None = None
     pass
 
@@ -26,13 +27,13 @@ class InventoryItemUpdate(BaseModel):
     section_id: int | None = None
     description: str | None = None
     has_critical_stock: bool | None = None
-    critical_stock_quantity: int | None = None
+    critical_stock_quantity: Decimal | None = None
     comments: str | None = None
 
 class InventoryItemRead(InventoryItemBase):
-    total_entries: int
-    total_exits: int
-    current_stock: int
+    total_entries: Decimal
+    total_exits: Decimal
+    current_stock: Decimal
     section_name: str
     deleted_at: datetime | None = None
