+++ Backend/internal/schemas/report_inventory_item_schema.py
@@ -1,9 +1,10 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 class ReportInventoryItemBase(BaseModel):
     report_id: int
     inventory_item_id: int
-    stock_at_generation: int
+    stock_at_generation: Decimal
 
 class ReportInventoryItemCreate(ReportInventoryItemBase):
     pass
